Implement Drop, figure out number limits
Now we close the read/write FDs when dropping the ring buffer (uh,
yeah!), which lets us create a whole bunch of those. Also, our tests now
determine how many files can be opened in addition to the one stored in
a ring buffer!

This is pretty amazing news - the test manages to open 128 more files
than =ulimit -a= says should be allowed!